Add --float-16 to llama-cli and associated support to llama-rs
This change adds a --float-16 flag which allows using a 16 bit float for the model memory
key and value tensors. The --float-16 flag enables it for both K and V, however internally
it is possible to set it independently (I don't know why you'd want to, but...)
In addition to saving memory, this change only reduces the size of the cached prompt by
about 50% when enabled.