Add --bias-tokens and --dump-prompt-tokens flags and associated backend changes.
--bias-tokens allows passing a list of token ids and float point biases. This can
be used to emulate llama.cpp's --ignore-eos feature in a more flexible way. For
example using "--bias-tokens 2=-1" will prevent the EOD token from being generated.

--dump-prompt-tokens will just dump the tokens for the prompt after loading, in two
formats: First just as a list of token ids and then as a list with the associated
string the token came from.

Some backend changes were necessary to make this possible but the behavior of the
existing code should remain unchanged. One thing I did do was move the tokenize
method to Vocabulary and changed it to generate Vec<(&str, TokenId)>. The
tokenize method in session remains as a wrapper to that which just strips out
the string part.