Added support for cuBLAS and CLBlast in ggml.
To enable either of them use --features cublas or --features clblast
when running cargo.

It looks like something went wrong yesterday with the latest commit in ggml and
ggml-opencl.* is kind of broken. So clblast in llm won't really work.
With older version of ggml, cblast in llm works.

By enabling the features only the prompt/encoding part will be faster.

It should be working with Ubuntu LTS 22.04 too. I tested it only with
Arch.